Split TestContextImplementation into focused partial-class files
TestContextImplementation.cs had grown to 988 lines and mixed several
unrelated concerns. The class is already partial, so the file is split
by responsibility with no API, behavior or import changes:

- TestContextImplementation.cs (242): ctor-injected state, constructors,
  core TestContext overrides, test-state setters, Dispose, clone
- TestContextImplementation.Output.cs (112): Write/WriteLine,
  DisplayMessage, diagnostic messages, SynchronizedStringBuilder accessors
- TestContextImplementation.Properties.cs (193): property bag, lifecycle
  property capture/merge, result files
- TestContextImplementation.TempDirectory.cs (222): per-test temp
  directory state, TestTempDirectory property, creation
- TestContextImplementation.TempDirectoryNaming.cs (117): length budget,
  name source, sanitization
- TestContextImplementation.TempDirectoryCleanup.cs (150): cleanup and
  retention policy

The temp-directory partials are guarded in their entirety by
'#if !WINDOWS_UWP && !WIN_UI', matching the guards that previously
wrapped those members.

Fixes #10270

Copilot-Session: 0509bdb6-c18d-4187-b463-7bf50769b9ea
